elbv2,website: mark default_action.target_group_arn as deprecated
The `target_group_arn` argument has been marked deprecated to remove
one of the ways to specify a target group to forward traffic to.
The `forward` block is now the preferred way to do it.

In case of one target group, it doesn't matter what arguments are
specified in the configuration, the ELB API will populate both
the `TargetGroupArn` and `ForwardConfig` parameters in the response.
The provider can't save both parameters to the state because there
would be a difference between the state and the configuration (where
only one argument is specified). So now `target_group_arn` is set from
the response only if it was specified in the configuration. Otherwise,
`forward` is set (also on import).